Finalized the first version of the neural networks
