+++ dtaianomaly/anomaly_detection/MultilayerPerceptron.py
@@ -1,6 +1,5 @@
-from typing import Literal
+from typing import Callable, Literal
 
-import numpy as np
 import torch
 
 from dtaianomaly import utils
@@ -8,15 +7,92 @@ from dtaianomaly.anomaly_detection.BaseDetector import Supervision
 from dtaianomaly.anomaly_detection.BaseNeuralDetector import (
     _ACTIVATION_FUNCTION_TYPE,
     _COMPILE_MODE_TYPE,
+    _MODEL_PARAMETERS_TYPE,
     _OPTIMIZER_TYPE,
-    BaseNeuralDetector,
-    ForecastDataset,
+)
+from dtaianomaly.anomaly_detection.BaseNeuralDetector_utils import (
+    BaseNeuralForecastingDetector,
 )
 
 
-class MultilayerPerceptron(BaseNeuralDetector):
+class MultilayerPerceptron(BaseNeuralForecastingDetector):
     """
-    TODO ...
+    Use a multilayer perceptron to detect anomalies.
+
+    The multilayer perceptron is a fully connected neural network which
+    will detect anomalies based on forecasting. Given a subsequence in the
+    time series, the network will learn to forecast the future values. Because
+    anomalies are unexpected events, they are difficult to forecast. Hence,
+    by computing the difference between the forecasted value and the actually
+    observed values, the neural network can detect anomalies.
+
+    The architecture of the multilayer perceptron consists of blocks, in which each
+    block applies the following operations: fully-connected layer :math:`\\rightarrow`
+    batch normalization :math:`\\rightarrow` activation function :math:`\\rightarrow`
+    dropout layer. The first and final layers of the network has no batch normalization,
+    the final layer of the network has no dropout.
+
+    Parameters
+    ----------
+    window_size: int or str
+        The window size to use for extracting sliding windows from the time series. This
+        value will be passed to :py:meth:`~dtaianomaly.anomaly_detection.compute_window_size`.
+    error_metric: {"mean-absolute-error", "mean-squared-error"}, default="mean-absolute-error"
+        The error measure between the reconstructed window and the original window.
+    hidden_layers: list of ints, default=[64, 32]
+        The number of neurons in each hidden layer. If an empty list is given, then the input
+        layer is directly connected to the output layer.
+    dropout_rate: float in interval [0, 1[, default=0.2
+        The dropout rate for the dropout layers. If the dropout rate is 0, no dropout layers
+        will be added to the auto encoder.
+    activation_function: {"linear", "relu", "sigmoid", "tanh"} default="relu"
+        The activation function to use at the end of each layer.
+    batch_normalization: bool = True,
+        Whether to add batch normalization after each layer or not.
+    stride: int, default=1
+        The stride, i.e., the step size for extracting sliding windows from the time series.
+    standard_scaling: bool, default=True
+        Whether to standard scale each window independently, before feeding it to the network.
+    batch_size: int, default=32
+        The size of the batches to feed to the network.
+    data_loader_kwargs: dictionary, default=None
+        Additional kwargs to be passed to the data loader.
+        For more information, see: https://docs.pytorch.org/docs/stable/data.html
+    optimizer: {"adam", "sgd"} or callable default="adam"
+        The optimizer to use for learning the weights. If "adam" is given,
+        then the torch.optim.Adam optimizer will be used. If "sgd" is given,
+        then the torch.optim.SGD optimizer will be used. Otherwise, a callable
+        should be given, which takes as input the network parameters, and then
+        creates an optimizer.
+    learning_rate: float, default=1e-3
+        The learning rate to use for training the network. Has no effect
+        if optimize is a callable.
+    compile_model: bool, default=False
+        Whether the network architecture should be compiled or not before
+        training the weights.
+        For more information, see: https://docs.pytorch.org/docs/stable/generated/torch.compile.html
+    compile_mode: {"default", "reduce-overhead", "max-autotune", "max-autotune-no-cudagraphs"}, default="default"
+        Method to compile the architecture.
+        For more information, see: https://docs.pytorch.org/docs/stable/generated/torch.compile.html
+    n_epochs: int, default=10
+        The number of epochs for which the neural network should be trained.
+    loss_function: torch.nn.Module, default=torch.nn.MSELoss()
+        The loss function to use for updating the weights.
+    device: str, default="cpu"
+        The device on which te neural network should be trained.
+        For more information, see: https://docs.pytorch.org/docs/stable/tensor_attributes.html#torch-device
+    seed: int, default=None
+        The seed used for training the model. This seed will update the torch
+        and numpy seed at the beginning of the fit method.
+
+    Attributes
+    ----------
+    window_size_: int
+        The effectively used window size for this anomaly detector.
+    optimizer_: torch.optim.Optimizer
+        The optimizer used for learning the weights of the network.
+    neural_network_: torch.nn.Module
+        The PyTorch network architecture.
 
     Examples
     --------
@@ -27,10 +103,14 @@ class MultilayerPerceptron(BaseNeuralDetector):
     >>> mlp.decision_function(x)  # doctest: +ELLIPSIS, +NORMALIZE_WHITESPACE
     array([1.8944391 , 1.8944391 , 1.83804671, ..., 0.59621549, 0.54421651,
            0.05852008]...)
+
+    See also
+    --------
+    BaseNeuralReconstructionDetector: Use a neural network to reconstruct
+        windows in the time series, and detect anomalies as windows that
+        are incorrectly reconstructed.
     """
 
-    error_metric: Literal["mean-absolute-error", "mean-squared-error"]
-    forecast_length: int
     hidden_layers: list[int]
     dropout_rate: float
     activation_function: _ACTIVATION_FUNCTION_TYPE
@@ -51,9 +131,10 @@ class MultilayerPerceptron(BaseNeuralDetector):
         standard_scaling: bool = True,
         batch_size: int = 32,
         data_loader_kwargs: dict[str, any] = None,
-        optimizer: _OPTIMIZER_TYPE = "adam",
+        optimizer: (
+            _OPTIMIZER_TYPE | Callable[[_MODEL_PARAMETERS_TYPE], torch.optim.Optimizer]
+        ) = "adam",
         learning_rate: float = 1e-3,
-        optimizer_kwargs: dict[str, any] = None,
         compile_model: bool = False,
         compile_mode: _COMPILE_MODE_TYPE = "default",
         n_epochs: int = 10,
@@ -62,15 +143,16 @@ class MultilayerPerceptron(BaseNeuralDetector):
         seed: int = None,
     ):
         super().__init__(
-            supervision=Supervision.SEMI_SUPERVISED,
             window_size=window_size,
+            supervision=Supervision.SEMI_SUPERVISED,
+            error_metric=error_metric,
+            forecast_length=forecast_length,
             stride=stride,
             standard_scaling=standard_scaling,
             batch_size=batch_size,
             data_loader_kwargs=data_loader_kwargs,
             optimizer=optimizer,
             learning_rate=learning_rate,
-            optimizer_kwargs=optimizer_kwargs,
             compile_model=compile_model,
             compile_mode=compile_mode,
             n_epochs=n_epochs,
@@ -78,18 +160,6 @@ class MultilayerPerceptron(BaseNeuralDetector):
             device=device,
             seed=seed,
         )
-
-        if not isinstance(error_metric, str):
-            raise TypeError("`error_metric` should be a string")
-        if error_metric not in ["mean-absolute-error", "mean-squared-error"]:
-            raise ValueError(
-                f"Unknown error_metric '{error_metric}'. Valid options are ['mean-absolute-error', 'mean-squared-error']"
-            )
-
-        if not isinstance(forecast_length, int) or isinstance(forecast_length, bool):
-            raise TypeError("`forecast_length` should be an integer")
-        if forecast_length < 1:
-            raise ValueError("`forecast_length` should be strictly positive")
 
         if not utils.is_valid_list(hidden_layers, int):
             raise TypeError("`hidden_layers` should be a list of integer")
@@ -102,7 +172,7 @@ class MultilayerPerceptron(BaseNeuralDetector):
             raise TypeError("`activation_function` should be a string")
         if activation_function not in self._ACTIVATION_FUNCTIONS:
             raise ValueError(
-                f"Unknown `activation_function` '{error_metric}'. Valid options are {list(self._ACTIVATION_FUNCTIONS.keys())}"
+                f"Unknown `activation_function` '{activation_function}'. Valid options are {list(self._ACTIVATION_FUNCTIONS.keys())}"
             )
 
         if not isinstance(batch_normalization, bool):
@@ -113,22 +183,10 @@ class MultilayerPerceptron(BaseNeuralDetector):
         if not 0.0 <= dropout_rate < 1.0:
             raise ValueError(f"`dropout_rate` should be in interval [0, 1[.")
 
-        self.error_metric = error_metric
-        self.forecast_length = forecast_length
         self.hidden_layers = hidden_layers
         self.dropout_rate = dropout_rate
         self.activation_function = activation_function
         self.batch_normalization = batch_normalization
-
-    def _build_dataset(self, X: np.ndarray) -> torch.utils.data.Dataset:
-        return ForecastDataset(
-            X=X,
-            window_size=self.window_size_,
-            stride=self.stride,
-            standard_scaling=self.standard_scaling,
-            device=self.device,
-            forecast_length=self.forecast_length,
-        )
 
     def _build_architecture(self, n_attributes: int) -> torch.nn.Module:
         # Initialize layer inputs and outputs
@@ -160,51 +218,3 @@ class MultilayerPerceptron(BaseNeuralDetector):
 
         # Return the encoder
         return mlp
-
-    def _train_batch(self, batch: list[torch.Tensor]) -> float:
-
-        # Set the type of the batch
-        history, future = batch
-
-        # Initialize the gradients to zero
-        self.optimizer_.zero_grad()
-
-        # Feed the data to the neural network
-        forecast = self.neural_network_(history)
-
-        # Compute the loss
-        loss = self.loss_function(forecast, future)
-
-        # Compute the gradients of the loss
-        loss.backward()
-
-        # Update the weights of the neural network
-        self.optimizer_.step()
-
-        # Return the loss
-        return loss.item()
-
-    def _evaluate_batch(self, batch: list[torch.Tensor]) -> torch.Tensor:
-
-        # Set the type of the batch
-        history, future = batch
-
-        # Forecast the data
-        forecast = self.neural_network_(history)
-
-        # Compute the difference with the given data
-        if self.error_metric == "mean-squared-error":
-            return torch.mean((forecast - future) ** 2, dim=1)
-        if self.error_metric == "mean-absolute-error":
-            return torch.mean(torch.abs(forecast - future), dim=1)
-
-        # Raise an error if invalid metric is given
-        raise ValueError(
-            f"Unknown error_metric '{self.error_metric}'. Valid options are ['mean-squared-error', 'mean-absolute-error']"
-        )
-
-    def _evaluate(self, data_loader: torch.utils.data.DataLoader) -> np.array:
-        decision_scores = super()._evaluate(data_loader)
-        return np.concatenate(
-            ([decision_scores[0] for _ in range(self.forecast_length)], decision_scores)
-        )
