+++ notebooks/README.md
@@ -1,0 +1,8 @@
+# More examples
+
+This directory a few examples on how ``dtaianomaly`` can be used
+to detect anomalies. You can find more examples in the 
+[online documentation](https://dtaianomaly.readthedocs.io/en/stable/getting_started/examples.html).
+
+The file [Config.json](./Config.json) gives a simple example of a 
+json-configuration file, which can be used to load a workflow.
